requirements.txt
Flask==3.1.0
flask-cors==5.0.1
flask-sock==0.7.0
openai==1.72.0
requests==2.32.3
python-dotenv==1.1.0
twilio==9.5.2
pydub==0.25.1
elevenlabs==0.2.27
gunicorn

